+++ app/Http/Livewire/Admin/Especificaciones/ShowEspecificaciones.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire\Admin\Especificaciones;
 use App\Models\Caracteristica;
 use App\Models\Especificacion;
 use App\Rules\CampoUnique;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Validation\Rule;
 use Livewire\Component;
@@ -13,7 +14,7 @@ use Livewire\WithPagination;
 class ShowEspecificaciones extends Component
 {
 
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     public $caracteristica;
     public $especificacion;
@@ -30,8 +31,8 @@ class ShowEspecificaciones extends Component
     {
         return [
             'caracteristica.name' => [
-                'required', 'min:3', 'max:100',
-                new CampoUnique('caracteristicas', 'name', $this->caracteristica->id, true),
+                'required', 'min:2', 'max:100',
+                new CampoUnique('caracteristicas', 'name', $this->caracteristica->id),
             ],
             'caracteristica.view' => [
                 'nullable', 'integer', 'min:0', 'max:1',
@@ -54,6 +55,7 @@ class ShowEspecificaciones extends Component
 
     public function edit(Caracteristica $caracteristica)
     {
+        $this->authorize('admin.almacen.caracteristicas.edit');
         $this->caracteristica = $caracteristica;
         $this->resetValidation();
         $this->open = true;
@@ -61,53 +63,48 @@ class ShowEspecificaciones extends Component
 
     public function update()
     {
+        $this->authorize('admin.almacen.caracteristicas.edit');
         $this->caracteristica->name = trim($this->caracteristica->name);
         $this->caracteristica->view = $this->caracteristica->view == 1 ? 1 : 0;
         $this->validate();
         $this->caracteristica->save();
         $this->reset(['open']);
+        $this->dispatchBrowserEvent('updated');
     }
 
-    public function openmodalespecificacion(Caracteristica $caracteristica)
+    public function addespecificacion(Caracteristica $caracteristica)
     {
-
-        $this->resetValidation(['name', 'especificacion']);
+        $this->authorize('admin.almacen.especificaciones.create');
+        $this->resetValidation();
         $this->reset(['name', 'especificacion']);
         $this->caracteristica = $caracteristica;
         $this->openespecificacion = true;
     }
 
-    public function add_especificacion()
+    public function saveespecificacion()
     {
+        $this->authorize('admin.almacen.especificaciones.create');
         $this->name = trim($this->name);
         $validatedata = $this->validate([
             'name' => [
                 'required', 'min:2',
                 Rule::unique('especificacions', 'name')
                     ->where('caracteristica_id', $this->caracteristica->id)
-                    ->whereNull('deleted_at')
             ]
         ]);
 
-
-        $especificacion = $this->caracteristica->especificacions()->withTrashed()
-            ->where('name', mb_strtoupper($this->name, "UTF-8"))->first();
-
-        if ($especificacion) {
-            $especificacion->restore();
-        } else {
-            $this->caracteristica->especificacions()->create([
-                'name' => $this->name
-            ]);
-        }
-
-        $this->resetValidation(['name']);
+        $this->caracteristica->especificacions()->create([
+            'name' => $this->name
+        ]);
+        $this->resetValidation();
         $this->reset(['name', 'openespecificacion']);
+        $this->dispatchBrowserEvent('created');
     }
 
     public function editespecificacion(Especificacion $especificacion)
     {
-        $this->resetValidation(['name', 'especificacion']);
+        $this->authorize('admin.almacen.especificaciones.edit');
+        $this->resetValidation();
         $this->reset(['name', 'especificacion']);
         $this->especificacion = $especificacion;
         $this->name = $especificacion->name;
@@ -116,6 +113,8 @@ class ShowEspecificaciones extends Component
 
     public function update_especificacion()
     {
+
+        $this->authorize('admin.almacen.especificaciones.edit');
         $this->name = trim($this->name);
         $validatedata = $this->validate([
             'name' => [
@@ -131,17 +130,19 @@ class ShowEspecificaciones extends Component
 
         $this->especificacion->name = trim($this->name);
         $this->especificacion->save();
-        $this->resetValidation(['name', 'especificacion']);
+        $this->resetValidation();
         $this->reset(['name', 'especificacion', 'openeditespecificacion']);
+        $this->dispatchBrowserEvent('updated');
     }
 
     public function delete(Caracteristica $caracteristica)
     {
+
+        $this->authorize('admin.almacen.caracteristicas.delete');
         DB::beginTransaction();
         try {
-
             $caracteristica->especificacions()->delete();
-            $caracteristica->deleteOrFail();
+            $caracteristica->delete();
             DB::commit();
             $this->dispatchBrowserEvent('deleted');
         } catch (\Exception $e) {
@@ -155,9 +156,11 @@ class ShowEspecificaciones extends Component
 
     public function deleteEspecificacion(Especificacion $especificacion)
     {
+
+        $this->authorize('admin.almacen.especificaciones.delete');
         DB::beginTransaction();
         try {
-            $especificacion->deleteOrFail();
+            $especificacion->delete();
             DB::commit();
             $this->dispatchBrowserEvent('deleted');
         } catch (\Exception $e) {
